requirements.txt
flask
flask_oidc
peewee
pymysql
python-decouple
